authorize: only redirect for HTML pages
